Indemnification
Bring back a little more legal language (with TIPs for non-lawyers) to constrain indemnification to the council and officers while giving the CFC discretion in extending that coverage to others (employees, committee members, agents).. without requiring a revision to the bylaws to do so.

Clarify that advance payments for indemnifications must be made in writing.